Update README.md
As our project is mainly a scenario based project here we can change the mode from day to night. Where we have some house, mountains, tree, grass, sun, and moon with stars. When we execute the program, we can find simple house and tree sketched with line. There will be mouse interaction which will let swap between the day and night view. There will be three mouse interaction added with use of menu. The three menu and their working will like,
1.The run the program and floating clouds.
2.Night color change &
3.Day color change.
These menu will come when we will press the mouse click. In day mode we will see sun and cloud while in night mode stars and moon. House, Mountain and Tree will be same in both the mode.